# server/user-service/src/main/kotlin/org/example/userservice/dto/response/ErrorResponse.kt
package org.example.userservice.dto.response

import java.time.LocalDateTime

data class ErrorResponse(
    val status: Int,
    val error: String,
    val message: String,
    val timestamp: LocalDateTime = LocalDateTime.now()
)